Sprint22_part5

+++ app/src/main/java/com/practicum/playlistmaker/media/data/db/PlaylistEntity.kt
@@ -11,6 +11,6 @@ data class PlaylistEntity(
     val playlistDescription: String?,
     val imagePath: String?,
     val listIdTracks: String = "",
-    val numberOfTracks: Long,
+    val numberOfTracks: Int,
     val timestamp: Long = System.currentTimeMillis()
 )
